requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
idna==3.11
joblib==1.5.2
multidict==6.7.0
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tinydb==4.8.2
typing_extensions==4.15.0
tzdata==2025.2
yarl==1.22.0
